dictionary prompt and history
Store history of dictionary lookups.

Create a function which prompts for words and looks them up in the same language
repeatedly.

Together these changes are meant to make it easier to watch Netflix in a foreign
language and look up words repeatedly without too many keystrokes getting in the way.
To improve it further, we may create a custom lesskey file to quit on any letter
input.